requirements.txt
tiktoken
llama-index
python-dotenv
pinecone-client
pymupdf
-e .